[feat] 메인 페이지 카드 슬라이드 1-1, 1-2 구현 완성
