+++ db.json
@@ -21,7 +21,7 @@
       "id": 1,
       "userId": 1,
       "amount": 3000000,
-      "date": "2025-04-01",
+      "date": "2025-03-01",
       "category": "월급",
       "description": "3월 월급",
       "payment": "이체",
