+++ src/pages/LJJ/MainPage.vue
@@ -1,7 +1,6 @@
 <script setup>
 import CategoryCard from '@/components/CategoryCard.vue'
-import ReportCard2 from '@/components/ReportCard2.vue'
-// import ReportCard from '@/components/ReportCard.vue'
+import ReportCard from '@/components/ReportCard.vue'
 </script>
 
 <template>
@@ -12,9 +11,7 @@ import ReportCard2 from '@/components/ReportCard2.vue'
   </div>
   <!--  -->
   <h6>main_slide_1_2 원본</h6>
-  <!-- <div><ReportCard /></div> -->
-  <h6>main_slide_1_2 수정중</h6>
-  <div><ReportCard2 /></div>
+  <div><ReportCard /></div>
 </template>
 
 <style scoped></style>
